Move the rest of the functionality from platform.h into core.
This also deprecates supplemental/util/platform.h.
