+++ docs/man/nng_cv_alloc.3supp.adoc
@@ -1,6 +1,6 @@
 = nng_cv_alloc(3supp)
 //
-// Copyright 2018 Staysail Systems, Inc. <info@staysail.tech>
+// Copyright 2024 Staysail Systems, Inc. <info@staysail.tech>
 // Copyright 2018 Capitar IT Group BV <info@capitar.com>
 //
 // This document is supplied under the terms of the MIT License, a
@@ -18,7 +18,6 @@ nng_cv_alloc - allocate condition variable
 [source, c]
 ----
 #include <nng/nng.h>
-#include <nng/supplemental/util/platform.h>
 
 typedef struct nng_cv nng_cv;
