+++ docs/man/nng_mtx_alloc.3supp.adoc
@@ -1,6 +1,6 @@
 = nng_mtx_alloc(3supp)
 //
-// Copyright 2018 Staysail Systems, Inc. <info@staysail.tech>
+// Copyright 2024 Staysail Systems, Inc. <info@staysail.tech>
 // Copyright 2018 Capitar IT Group BV <info@capitar.com>
 //
 // This document is supplied under the terms of the MIT License, a
@@ -18,7 +18,6 @@ nng_mtx_alloc - allocate mutex
 [source, c]
 ----
 #include <nng/nng.h>
-#include <nng/supplemental/util/platform.h>
 
 typedef struct nng_mtx nng_mtx;
