+++ src/supplemental/websocket/websocket_test.c
@@ -1,5 +1,5 @@
 //
-// Copyright 2020 Staysail Systems, Inc. <info@staysail.tech>
+// Copyright 2024 Staysail Systems, Inc. <info@staysail.tech>
 // Copyright 2018 Capitar IT Group BV <info@capitar.com>
 //
 // This software is supplied under the terms of the MIT License, a
@@ -10,7 +10,6 @@
 #include <string.h>
 
 #include <nng/nng.h>
-#include <nng/supplemental/util/platform.h>
 
 #include <nuts.h>
