+++ tests/synch.c
@@ -1,6 +1,6 @@
 //
 // Copyright 2022 Staysail Systems, Inc. <info@staysail.tech>
-// Copyright 2018 Capitar IT Group BV <info@capitar.com>
+// Copyright 2024 Capitar IT Group BV <info@capitar.com>
 //
 // This software is supplied under the terms of the MIT License, a
 // copy of which should be located in the distribution where this
@@ -9,7 +9,6 @@
 //
 
 #include <nng/nng.h>
-#include <nng/supplemental/util/platform.h>
 
 #include "convey.h"
 
@@ -17,8 +16,8 @@
 struct notifyarg {
 	int          did;
 	nng_duration when;
-	nng_mtx *    mx;
-	nng_cv *     cv;
+	nng_mtx     *mx;
+	nng_cv      *cv;
 };
 
 void
@@ -34,7 +33,7 @@ notifyafter(void *a)
 }
 
 struct notifyarg arg;
-nng_thread *     thr;
+nng_thread      *thr;
 
 static void
 test_sync(void)
